+++ shaders/import/nightVision.glsl
@@ -1,0 +1,5 @@
+
+#define import_nightVision
+#ifdef FIRST_PASS
+float nightVision = 0.0;
+#endif