+++ shaders/main_files/gbuffers_clouds.glsl
@@ -72,7 +72,7 @@ void main() {
 	texcoord = (gl_TextureMatrix[0] * gl_MultiTexCoord0).xy;
 	
 	vec3 skyLight = getSkyLight(ARG_IN);
-	vec3 ambientLight = getAmbientLight(ARG_IN);
+	vec3 ambientLight = getAmbientLight(1.0  ARGS_IN);
 	colorMult = skyLight + ambientLight;
 	//colorMult = mix(vec3(getColorLum(colorMult)), colorMult, vec3(1.0));
 	colorMult = normalize(colorMult) * 2.0 * CLOUDS_BRIGHTNESS;